feat(engine): rpc abi handlers fall back via codehash on miss
`edb_getContractABI`, `edb_getCallableABI`, and `edb_getConstructorArgs`
now consult the `codehash_to_canonical` index when the queried address
has no direct entry in `artifacts` / `recompiled_artifacts`. This
completes the UI-facing half of the etched-contract source-view fix:
every artifact-address-keyed RPC handler now resolves `vm.etch`-
aliased addresses to their canonical artifact, so ABI/source/ctor-args
views render correctly for an address whose live bytecode matches a
known recompiled runtime.

Also promoted `bytecode_at` and `resolve_canonical_via_codehash` from
private helpers in `rpc/methods/artifact.rs` to `pub(crate)` methods on
`EngineContext`. The two helpers were already needed by both the
`edb_getCode*` family (T4) and this commit's ABI handlers; hoisting
them to the context makes the API discoverable from any RPC handler
holding an `&EngineContext`. Behavior of the existing T4 code paths is
identical — the methods carry the same trace walk, empty-guard, hash,
and lookup, only the call sites changed.

Sweep notes: analysis-results lookups in `breakpoint.rs`,
`navigation.rs`, and `snapshot.rs` are keyed by the etched
`bytecode_address` and would benefit from a parallel fallback, but
they target the `analysis_results` map (not `artifacts`) and need a
slightly different design (resolve once upstream so the analysis,
step, and source lookups all share the canonical key). Left for a
follow-up.
