+++ crates/engine/src/rpc/methods/resolve.rs
@@ -57,10 +57,20 @@ where
             data: None,
         })?;
 
+    // Resolve via the direct address first; if that misses, fall back
+    // through the codehash alias index so addresses installed by
+    // `vm.etch(target, address(impl).code)` surface the canonical
+    // artifact's ABI instead of `null`. The same fallback already
+    // backs `edb_getCode` (see `artifact.rs`) and the hook inspector
+    // (see `HookSnapshotInspector::resolve_analysis`).
     let abi = if recompiled {
         context
             .recompiled_artifacts
             .get(&address)
+            .or_else(|| {
+                let canonical = context.resolve_canonical_via_codehash(address)?;
+                context.recompiled_artifacts.get(&canonical)
+            })
             .and_then(|artifact| artifact.contract())
             .and_then(|contract| contract.abi.as_ref())
             .cloned()
@@ -68,6 +78,10 @@ where
         context
             .artifacts
             .get(&address)
+            .or_else(|| {
+                let canonical = context.resolve_canonical_via_codehash(address)?;
+                context.artifacts.get(&canonical)
+            })
             .and_then(|artifact| artifact.contract())
             .and_then(|contract| contract.abi.as_ref())
             .cloned()
@@ -142,9 +156,22 @@ where
     let abi_info = related_addresses
         .into_iter()
         .filter_map(|(addr, ty)| {
-            context.recompiled_artifacts.get(&addr).and_then(|artifact| {
-                artifact.contract().map(|contract| parse_callable_abi_info(addr, contract, ty))
-            })
+            // Direct lookup first; fall back through the codehash alias
+            // index so a related address whose code was installed by
+            // `vm.etch` still resolves to the canonical recompiled
+            // artifact. The `addr` returned to the UI is preserved
+            // (not the canonical one) so callers can target the live
+            // address they actually queried.
+            context
+                .recompiled_artifacts
+                .get(&addr)
+                .or_else(|| {
+                    let canonical = context.resolve_canonical_via_codehash(addr)?;
+                    context.recompiled_artifacts.get(&canonical)
+                })
+                .and_then(|artifact| {
+                    artifact.contract().map(|contract| parse_callable_abi_info(addr, contract, ty))
+                })
         })
         .collect::<Vec<CallableAbiInfo>>();
